data_processing/requirements.txt
pyspark==3.5.1
kafka-python
scikit-learn
pandas
numpy
